streamObject
