+++ packages/core/core/generate-object/index.ts
@@ -1,3 +1,4 @@
 export * from './generate-object';
 export * from './generate-object-result';
 export * from './stream-object';
+export * from './stream-object-result';
